Create API-Delete savings

+++ src/main/java/com/lali/financial/habits/management/service/repository/SavingsRepository.java
@@ -39,6 +39,6 @@ public interface SavingsRepository extends JpaRepository<Savings, Long> {
     @Modifying
     @Transactional
     @Query("delete from Savings a where a.savingsId=?1")
-    void deleteBySavingsId(Integer savingsId);
+    void deleteBySavingsId(Long savingsId);
 
 }
